bai cua Tri

+++ 132/132.cpp
@@ -28,10 +28,10 @@ int main()
 	cout << "Nhap ym: ";
 	cin >> ym;
 
-	float sabc = (1 / 2) * (xa * yb + xb * yc + xc * ya - xb * ya - xc * yb - xa * yc);
-	float smac = (1 / 2) * (xm * ya + xa * yc + xc * ym - xa * ym - xc * ya - xm * yc);
-	float smbc = (1 / 2) * (xm * yb + xb * yc + xc * ym - xb * ym - xc * yb - xm * yc);
-	float smab = (1 / 2) * (xm * ya + xa * yb + xb * ym - xa * ym - xb * ya - xm * yb);
+	float sabc = (float)(1 / 2) * (xa * yb + xb * yc + xc * ya - xb * ya - xc * yb - xa * yc);
+	float smac = (float)(1 / 2) * (xm * ya + xa * yc + xc * ym - xa * ym - xc * ya - xm * yc);
+	float smbc = (float)(1 / 2) * (xm * yb + xb * yc + xc * ym - xb * ym - xc * yb - xm * yc);
+	float smab = (float)(1 / 2) * (xm * ya + xa * yb + xb * ym - xa * ym - xb * ya - xm * yb);
 
 	if (sabc == (smac + smab + smbc))
 		cout << "Nam trong tam giac";
